# movie-booking-api-spec.yaml
openapi: 3.1.0
info:
  title: Movie Booking API
  version: 1.2.0
  description: >
    REST API for browsing movies, viewing showtimes, reserving seats, and
    managing bookings.

    All resource IDs are strings (UUIDs). All timestamps are ISO 8601 UTC
    (e.g. 2026-04-05T10:00:00Z).

    Field names use camelCase throughout to match the frontend TypeScript
    types directly (no snake_case conversion required).
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                email:
                  type: string
                  format: email
                  maxLength: 255
                password:
                  type: string
                  minLength: 8
                  maxLength: 72
                  description: Plain-text password; hashed server-side before persistence.
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/EmailConflict'
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate with email and password
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
                password:
                  type: string
                  minLength: 8
                  maxLength: 72
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Invalidate current refresh token
      operationId: logoutUser
      responses:
        '204':
          description: Logged out successfully
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Users ─────────────────────────────────────────────────────────────────────
  /users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      operationId: getMyProfile
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/bookings:
    get:
      tags: [Users]
      summary: List my bookings (ticket history)
      operationId: listMyBookings
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: status
          in: query
          required: false
          description: Filter by booking status.
          schema:
            $ref: '#/components/schemas/BookingStatus'
      responses:
        '200':
          description: >
            Paginated booking list. Each item embeds the showtime and movie
            so the tickets page can render without additional fetches.
          content:
            application/json:
              schema:
                type: object
                required: [items, page, size, totalItems, totalPages]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/BookingWithContext'
                  page:
                    type: integer
                    minimum: 0
                  size:
                    type: integer
                    minimum: 1
                  totalItems:
                    type: integer
                    minimum: 0
                  totalPages:
                    type: integer
                    minimum: 0
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Movies ────────────────────────────────────────────────────────────────────
...
  /showtimes/{showtimeId}:
    get:
      tags: [Showtimes]
      summary: Get a single showtime by ID
      operationId: getShowtimeById
      security: []
      description: >
        Returns showtime details used by the booking page header (room name,
        start time, price). Call this in parallel with
        GET /showtimes/{showtimeId}/seats to populate the full booking view.
      parameters:
        - $ref: '#/components/parameters/ShowtimeIdPath'
      responses:
        '200':
          description: Showtime detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShowtimeItem'
        '404':
          $ref: '#/components/responses/ShowtimeNotFound'
...
  /admin/showtimes:
    get:
      tags: [Admin Showtimes]
      summary: List all showtimes
      operationId: adminListShowtimes
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: movieId
          in: query
          required: false
          schema:
            type: string
        - name: fromTime
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: toTime
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: >
            Paginated showtime list. Each item includes movie title and
            cinema/room details needed by the admin scheduling panel.
          content:
            application/json:
              schema:
                type: object
                required: [items, page, size, totalItems, totalPages]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminShowtimeItem'
                  page:
                    type: integer
                    minimum: 0
                  size:
                    type: integer
                    minimum: 1
                  totalItems:
                    type: integer
                    minimum: 0
                  totalPages:
                    type: integer
                    minimum: 0
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Admin Showtimes]
      summary: Create a showtime
      operationId: createShowtime
      description: >
        The end time is derived from the movie's durationMinutes.
        The room capacity is derived from the selected room.
        Overlap check is performed against existing showtimes in the same room.
        Adjacent showtimes are allowed: a showtime may start exactly when the
        previous one ends (startTime == previousEndTime is not an overlap).
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShowtimeWriteBody'
      responses:
        '201':
          description: Showtime created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminShowtimeItem'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/MovieNotFound'
        '409':
          $ref: '#/components/responses/ShowtimeOverlap'
  /admin/showtimes/{showtimeId}:
    patch:
      tags: [Admin Showtimes]
      summary: Update a showtime
      operationId: updateShowtime
      parameters:
        - $ref: '#/components/parameters/ShowtimeIdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShowtimeWriteBody'
      responses:
        '200':
          description: Showtime updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminShowtimeItem'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ShowtimeNotFound'
        '409':
          $ref: '#/components/responses/ShowtimeOverlap'

    delete:
      tags: [Admin Showtimes]
      summary: Delete a showtime
      operationId: deleteShowtime
      parameters:
        - $ref: '#/components/parameters/ShowtimeIdPath'
      responses:
        '204':
          description: Showtime deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ShowtimeNotFound'

  # ── Admin Cinemas ─────────────────────────────────────────────────────────────
...
  /bookings:
    post:
      tags: [Bookings]
      summary: Create a booking
      operationId: createBooking
      description: >
        Creates a PENDING booking and marks the selected seats as HELD.
        The booking must be confirmed via PATCH /bookings/{bookingId}/confirm
        to become CONFIRMED. Seats held by expired PENDING bookings are
        released automatically by the server.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [showtimeId, seatIds]
              properties:
                showtimeId:
                  type: string
                seatIds:
                  type: array
                  minItems: 1
                  maxItems: 6
                  uniqueItems: true
                  items:
                    type: string
                  description: >
                    IDs of seats from GET /showtimes/{showtimeId}/seats.
                    Maximum 6 seats per booking.
      responses:
        '201':
          description: Booking created (PENDING); seats are held.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingItem'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ShowtimeNotFound'
        '409':
          $ref: '#/components/responses/SeatConflict'
  /bookings/{bookingId}:
    get:
      tags: [Bookings]
      summary: Get booking details
      operationId: getBooking
      description: >
        Returns the full booking including seat objects.
        Used by the checkout page to display booking summary before confirmation.
      parameters:
        - $ref: '#/components/parameters/BookingIdPath'
      responses:
        '200':
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/BookingNotFound'

    patch:
      tags: [Bookings]
      summary: Cancel a booking
      operationId: cancelBooking
      description: >
        Transitions the booking to CANCELLED and releases all held or booked
        seats back to AVAILABLE.
      parameters:
        - $ref: '#/components/parameters/BookingIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              description: Only CANCELLED is accepted.
              properties:
                status:
                  type: string
                  enum: [CANCELLED]
      responses:
        '200':
          description: Booking cancelled; seats released.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingItem'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/BookingNotFound'
        '409':
          $ref: '#/components/responses/BookingAlreadyCancelled'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ShowtimeIdPath:
      name: showtimeId
      in: path
      required: true
      schema:
        type: string
    BookingIdPath:
      name: bookingId
      in: path
      required: true
      schema:
        type: string
    PageParam:
      name: page
      in: query
      required: false
      description: Zero-based page index.
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      description: Number of items per page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      required: false
      description: |
        Sort order as `field:direction`. Direction must be `asc` or `desc`.
        Example: `createdAt:desc`.
      schema:
        type: string
        pattern: '^[a-zA-Z]+:(asc|desc)$'
        examples:
          - createdAt:desc
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Client-generated key for safe retries. Repeated requests with the same
        key return the original response without creating a duplicate resource.
      schema:
        type: string
        maxLength: 100
  schemas:
...
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, tokenType, expiresIn, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          examples:
            - Bearer
        expiresIn:
          type: integer
          format: int32
          description: Access token lifetime in seconds.
        user:
          $ref: '#/components/schemas/UserProfile'
    UserProfile:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        createdAt:
          type: string
          format: date-time

    # ── Movies ────────────────────────────────────────────────────────────────
...
    ShowtimeItem:
      type: object
      required: [id, movieId, roomName, startsAt, endsAt, price, capacity, remainingSeats, createdAt, updatedAt]
      properties:
        id:
          type: string
        movieId:
          type: string
        roomName:
          type: string
          description: Display name of the room (e.g. "Auditorium A").
          examples:
            - Auditorium A
        startsAt:
          type: string
          format: date-time
          description: UTC start time.
        endsAt:
          type: string
          format: date-time
          description: UTC end time (derived from movie durationMinutes).
        price:
          type: number
          format: float
          description: Ticket price for this showtime.
        capacity:
          type: integer
          description: Total seat count in the room.
        remainingSeats:
          type: integer
          description: Seats still available for booking (not HELD or BOOKED).
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminShowtimeItem:
      description: >
        Showtime with additional fields needed by the admin scheduling panel:
        movie title, duration, cinema/room identifiers and names, and lifecycle status.
      allOf:
        - $ref: '#/components/schemas/ShowtimeItem'
        - type: object
          required: [movieTitle, durationMinutes, cinemaId, cinemaName, roomId, status]
          properties:
            movieTitle:
              type: string
            durationMinutes:
              type: integer
              description: Runtime in minutes, copied from the movie. Used in the admin schedule card display.
            cinemaId:
              type: string
            cinemaName:
              type: string
            roomId:
              type: string
            status:
              $ref: '#/components/schemas/ShowtimeStatus'
    ShowtimeWriteBody:
      type: object
      required: [movieId, cinemaId, roomId, startsAt, price]
      description: >
        The end time is derived server-side from movie durationMinutes.
        The room capacity is derived from the selected room.
      properties:
        movieId:
          type: string
        cinemaId:
          type: string
        roomId:
          type: string
        startsAt:
          type: string
          format: date-time
          description: UTC start time (e.g. 2026-04-05T10:00:00Z).
        price:
          type: number
          format: float
          minimum: 0
          description: Ticket price for this showtime.

    # ── Seats ─────────────────────────────────────────────────────────────────
    SeatItem:
      type: object
      required: [id, label, row, number, status]
      properties:
        id:
          type: string
        label:
          type: string
          description: Combined row + number label (e.g. A1, B7).
          examples:
            - A1
        row:
          type: string
          description: Single uppercase letter identifying the row (A–Z).
          pattern: '^[A-Z]$'
          examples:
            - A
        number:
          type: integer
          minimum: 1
          description: Seat position within the row.
        status:
          $ref: '#/components/schemas/SeatStatus'

    # ── Cinemas ───────────────────────────────────────────────────────────────
...
    BookingItem:
      type: object
      required: [id, showtimeId, seatIds, seats, unitPrice, totalAmount, status, createdAt]
      properties:
        id:
          type: string
        showtimeId:
          type: string
        seatIds:
          type: array
          items:
            type: string
          description: IDs of the booked seats.
        seats:
          type: array
          items:
            $ref: '#/components/schemas/SeatItem'
          description: Full seat objects for display in checkout and ticket history.
        unitPrice:
          type: number
          format: float
          description: Price per seat at the time of booking.
        totalAmount:
          type: number
          format: float
          description: Total booking cost (unitPrice × seat count).
        status:
          $ref: '#/components/schemas/BookingStatus'
        createdAt:
          type: string
          format: date-time
        confirmedAt:
          type: string
          format: date-time
          nullable: true
          description: Set when status transitions to CONFIRMED.
    BookingWithContext:
      description: >
        Booking enriched with the showtime and movie it belongs to.
        Used by GET /users/me/bookings so the ticket history page can
        render without additional fetches.
      allOf:
        - $ref: '#/components/schemas/BookingItem'
        - type: object
          required: [showtime, movie]
          properties:
            showtime:
              type: object
              required: [id, startsAt, endsAt, roomName]
              properties:
                id:
                  type: string
                startsAt:
                  type: string
                  format: date-time
                endsAt:
                  type: string
                  format: date-time
                roomName:
                  type: string
            movie:
              type: object
              required: [id, title]
              properties:
                id:
                  type: string
                title:
                  type: string

    # ── Enums ─────────────────────────────────────────────────────────────────
    UserRole:
      type: string
      enum: [USER, ADMIN]
    ShowtimeStatus:
      type: string
      enum: [SCHEDULED, ONGOING, COMPLETED, CANCELLED]
      description: >
        Lifecycle state of a showtime.
        SCHEDULED: not yet started.
        ONGOING: currently in progress.
        COMPLETED: screening has ended.
        CANCELLED: showtime was cancelled; seats are released.
...
    BookingStatus:
      type: string
      enum: [PENDING, CONFIRMED, CANCELLED]
    SeatStatus:
      type: string
      enum: [AVAILABLE, HELD, BOOKED]

    # ── Errors ────────────────────────────────────────────────────────────────
    ProblemDetail:
      type: object
      required: [code, detail]
      properties:
        code:
          type: string
          description: Machine-readable error code.
          examples:
            - VALIDATION_ERROR
        detail:
          type: string
          description: Human-readable explanation of this occurrence.
          examples:
            - Request validation failed.
        errors:
          type: array
          description: Field-level errors. Present on 400 responses only.
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
                examples:
                  - email
              message:
                type: string
                examples:
                  - must be a well-formed email address
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            default:
              value:
                code: VALIDATION_ERROR
                detail: Request validation failed.
                errors:
                  - field: email
                    message: must be a well-formed email address
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            default:
              value:
                code: UNAUTHORIZED
                detail: Authentication is required.
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            default:
              value:
                code: FORBIDDEN
                detail: You do not have permission to access this resource.
    EmailConflict:
      description: Email is already registered
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            default:
              value:
                code: EMAIL_ALREADY_EXISTS
                detail: The email address is already registered.
                errors:
                  - field: email
                    message: already registered
    MovieNotFound:
      description: Movie not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            default:
              value:
                code: MOVIE_NOT_FOUND
                detail: No movie found with the given ID.
    ShowtimeNotFound:
      description: Showtime not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            default:
              value:
                code: SHOWTIME_NOT_FOUND
                detail: No showtime found with the given ID.
    BookingNotFound:
      description: Booking not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            default:
              value:
                code: BOOKING_NOT_FOUND
                detail: No booking found with the given ID.
    ShowtimeOverlap:
      description: Showtime overlaps an existing showtime in the same room
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            default:
              value:
                code: SHOWTIME_OVERLAP
                detail: The requested time slot overlaps an existing showtime in this room.
                errors:
                  - field: schedule
                    message: overlapping schedule
    SeatConflict:
      description: One or more seats are already reserved
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            default:
              value:
                code: SEAT_ALREADY_RESERVED
                detail: One or more selected seats are already reserved.
                errors:
                  - field: seatIds
                    message: already reserved
    BookingAlreadyCancelled:
      description: Booking is already cancelled
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            default:
              value:
                code: BOOKING_ALREADY_CANCELLED
                detail: This booking has already been cancelled.
...